// Frontend/quran-dashboard-ui/src/app/core/api/generated/models/linking-prepared-source-body.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { LinkingPreparedInlineSourceBody } from '../models/linking-prepared-inline-source-body';
import { LinkingPreparedWorkspaceSourceBody } from '../models/linking-prepared-workspace-source-body';
export interface LinkingPreparedSourceBody {
  inlineSource: LinkingPreparedInlineSourceBody | null;
  orderValue: number | null;
  workspaceSource: LinkingPreparedWorkspaceSourceBody | null;
}