// Frontend/quran-dashboard-ui/src/app/core/api/generated/models/create-linking-prepared-preflight-body.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { LinkingPreparedSourceBody } from '../models/linking-prepared-source-body';
export interface CreateLinkingPreparedPreflightBody {
  doorId: number | null;
  expectedLinkingDataRevision: number | null;
  preparationKey: string | null;
  sources: Array<LinkingPreparedSourceBody> | null;
}